requirements.txt
rpaframework==28.6.0
selenium==4.15.2
flake8==7.0.0
pandas==2.2.2
opencv-python==4.10.0.84
pyautogui==0.9.54